Upload started notification
